Bump version to 0.1.14; add LaTeX table export, expand keywords, refine XLS example string handling, and update documentation
Update version in Cargo.toml, AGENTS.md, and README.md. Add `latex` output format to `OutputFormat` enum with `print_latex()` implementations in `IoCommandHandler` and `PandasCommandHandler` — generates `\begin{tabular}{lll}\hline ... \end{tabular}` with header row, data rows, and horizontal rules. Expand Cargo.toml keywords to include `parquet`, `cli`, and `data-analysis

+++ src/excel/xls_writer/biff.rs
@@ -1086,7 +1086,7 @@ pub fn tab_id(sheet_count: u16) -> Vec<u8> {
     let mut out = Vec::new();
     let mut rec = Record::new(RecordId::TabId);
     for i in 0..sheet_count {
-        rec = rec.u16(i as u16 + 1);
+        rec = rec.u16(i + 1);
     }
     rec.finish(&mut out).unwrap();
     out
